requirements.txt
absl-py==0.8.1
argcomplete==1.11.1
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
bashplotlib==0.6.5
bleach==3.1.1
cachetools==3.1.1
certifi==2020.6.20
cffi==1.13.2
chardet==3.0.4
click==7.1.2
cloudpickle==1.2.2
colorama==0.4.3
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
defusedxml==0.6.0
dill==0.2.7.1
dm-env==1.2
dm-tree==0.1.5
docopt==0.6.2
dominate==2.5.1
entrypoints==0.3
fasteners==0.15
fn==0.4.3
future==0.18.2
fvcore==0.1.2.post20201020
gast==0.2.2
gitdb==4.0.5
GitPython==3.1.3
glfw==1.12.0
grpcio==1.25.0
gtimer==1.0.0b5
gym==0.10.9
h5py==2.10.0
idna==2.10
imageio==2.6.1
imageio-ffmpeg==0.3.0
importlib-metadata==1.5.0
imutils==0.5.3
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonpatch==1.26
jsonpickle==1.4.1
jsonpointer==2.0
jsonschema==3.2.0
kiwisolver==1.1.0
labmaze==1.0.2
llvmlite==0.30.0
lxml==4.5.2
Markdown==3.1.1
MarkupSafe==1.1.1
magnum==11.0.0
matplotlib==3.1.2
mistune==0.8.4
mock==3.0.5
monotonic==1.5
mpmath==1.1.0
msgpack==1.0.0
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
notebook==6.0.3
numba==0.46.0
numexpr==2.7.1
numpy==1.19.1
numpy-quaternion==2019.10.3.10.26.21
opencv-python==3.4.2.17
opt-einsum==3.1.0
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.1.0
plotly==4.4.1
portalocker==1.5.2
progressbar2==3.51.3
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.0rc2
ptyprocess==0.6.0
py-cpuinfo==6.0.0
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pycocotools-fix==2.0.0.1
PyFunctional==1.3.0
pyglet==1.5.0
Pygments==2.5.2
pymongo==3.11.0
PyOpenGL==3.1.5
pyparsing==2.4.5
pyrsistent==0.15.7
python-dateutil==2.8.1
python-utils==2.4.0
pytube==9.7.0
pytube3==9.6.4
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
sacred==0.8.1
scikit-fmm==2019.1.30
scikit-image==0.16.2
scikit-learn==0.22.2
scipy==1.5.2
Send2Trash==1.5.0
Shapely==1.7.0
six==1.15.0
sklearn==0.0
smmap==3.0.4
stable-baselines==2.9.0
sympy==1.5.1
tables==3.6.1
tabulate==0.8.6
tensorboard==2.1.0
tensorboardX==1.9
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torch==1.3.1
torchfile==0.1.0
torchvision==0.4.2
tornado==6.0.4
tqdm==4.39.0
traitlets==4.3.3
typing-extensions==3.7.4.2
urllib3==1.25.10
visdom==0.1.8.9
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
yacs==0.1.6
yapf==0.29.0
zipp==3.1.0
